old site assets deleted, index integrated with react
