+++ src/components/projectsDB.js
@@ -1,0 +1,23 @@
+'use strict'
+
+var ProjectsDB = [
+	{
+		name: 'Aprende y Crece',
+		img: '/img/projects/aprendeYCrece.jpg',
+		link: 'http://www.aprendeycrece.com.mx/',
+		tags: [
+			{type: 'title', content: 'Responsive Arquitecture'},
+			{type: 'title', content: 'Development'},
+			{type: 'regular', content: 'Responsive and liquid layout'},
+			{type: 'regular', content: 'Cross browser'},
+			{type: 'regular', content: 'Multiligual'},
+			{type: 'regular', content: 'E-learning platform'},
+			{type: 'regular', content: 'Social API'},
+			{type: 'regular', content: 'Accesibility'},
+			{type: 'regular', content: 'Dynamic site'},
+			{type: 'regular', content: 'Custom components'},
+		]
+	}
+];
+
+module.exports = ProjectsDB;
