+++ src/main.js
@@ -1,21 +1,19 @@
 $ = jQuery = require('jquery');
 var React = require('react');
 var ReactDOM = require('react-dom');
+var main = require('./js/base');
 
 // Components
-var Nav = require('./components/Nav');
-var Home = require('./components/Home');
-
+var Projects = require('./components/Projects');
 
 var App = React.createClass({
 	render: function(){
 		return (
-			<div id='wrapper'>
-				<Nav />
-				<Home />
+			<div id='itemsWrapper'>
+				
 			</div>
 		)
 	}
 });
 
-ReactDOM.render(<App />, document.getElementById('app'));
+ReactDOM.render(<App />, document.getElementById('projects'));
